Move reservation models to separate files
