+++ reservations/models/age_group.py
@@ -1,0 +1,17 @@
+from django.db import models
+
+__all__ = [
+    "AgeGroup",
+]
+
+
+class AgeGroup(models.Model):
+    minimum = models.fields.PositiveIntegerField(null=False, blank=False)
+    maximum = models.fields.PositiveIntegerField(null=True, blank=True)
+
+    class Meta:
+        db_table = "age_group"
+        base_manager_name = "objects"
+
+    def __str__(self) -> str:
+        return f"{self.minimum} - {self.maximum}"
