+++ reservations/models/reservation.py
@@ -1,8 +1,3 @@
-from datetime import date, datetime
-from uuid import UUID, uuid4
-
-from django.contrib.auth import get_user_model
-from django.core.validators import validate_comma_separated_integer_list
 from django.db import models
 from django.utils import timezone
 from django.utils.translation import gettext_lazy as _
@@ -10,9 +5,7 @@ from django_prometheus.models import ExportModelOperationsMixin
 from helsinki_gdpr.models import SerializableMixin
 
 from applications.choices import PriorityChoice
-from applications.models import City
 from common.connectors import ReservationActionsConnector
-from reservation_units.models import ReservationUnit
 from reservations.choices import (
     RESERVEE_LANGUAGE_CHOICES,
     CustomerTypeChoice,
@@ -21,134 +14,10 @@ from reservations.choices import (
 )
 from reservations.querysets import ReservationManager
 from tilavarauspalvelu.utils.auditlog_util import AuditLogger
-from tilavarauspalvelu.utils.commons import WEEKDAYS
 
-User = get_user_model()
-
-
-class AgeGroup(models.Model):
-    minimum = models.fields.PositiveIntegerField(null=False, blank=False)
-    maximum = models.fields.PositiveIntegerField(null=True, blank=True)
-
-    class Meta:
-        db_table = "age_group"
-        base_manager_name = "objects"
-
-    def __str__(self) -> str:
-        return f"{self.minimum} - {self.maximum}"
-
-
-class AbilityGroup(models.Model):
-    name = models.fields.TextField(null=False, blank=False, unique=True)
-
-    class Meta:
-        db_table = "ability_group"
-        base_manager_name = "objects"
-
-    def __str__(self) -> str:
-        return self.name
-
-
-class ReservationCancelReason(models.Model):
-    reason = models.CharField(max_length=255, null=False, blank=False)
-
-    class Meta:
-        db_table = "reservation_cancel_reason"
-        base_manager_name = "objects"
-
-    def __str__(self) -> str:
-        return self.reason
-
-
-class ReservationDenyReason(models.Model):
-    reason = models.CharField(max_length=255, null=False, blank=False)
-
-    class Meta:
-        db_table = "reservation_deny_reason"
-        base_manager_name = "objects"
-
-    def __str__(self) -> str:
-        return self.reason
-
-
-class RecurringReservation(models.Model):
-    name: str = models.CharField(max_length=255, blank=True, default="")
-    description: str = models.CharField(max_length=500, blank=True, default="")
-    uuid: UUID = models.UUIDField(default=uuid4, editable=False, unique=True)
-    user: User | None = models.ForeignKey(User, on_delete=models.SET_NULL, null=True)
-
-    begin_date: date | None = models.DateField(null=True)
-    begin_time: date | None = models.TimeField(null=True)
-    end_date: date | None = models.DateField(null=True)
-    end_time: date | None = models.TimeField(null=True)
-
-    application_event_schedule = models.ForeignKey(
-        "applications.ApplicationEventSchedule",
-        on_delete=models.PROTECT,
-        null=True,
-        blank=True,
-        related_name="recurring_reservations",
-    )
-
-    reservation_unit = models.ForeignKey(
-        "reservation_units.ReservationUnit",
-        on_delete=models.PROTECT,
-        related_name="recurring_reservations",
-    )
-
-    recurrence_in_days: int | None = models.PositiveIntegerField(null=True)
-    """How many days between reoccurring reservations"""
-
-    weekdays: str = models.CharField(
-        max_length=16,
-        validators=[validate_comma_separated_integer_list],
-        choices=WEEKDAYS.CHOICES,
-        blank=True,
-        default="",
-    )
-
-    age_group = models.ForeignKey(
-        "reservations.AgeGroup",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="recurring_reservations",
-    )
-
-    ability_group = models.ForeignKey(
-        "reservations.AbilityGroup",
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-        related_name="recurring_reservations",
-    )
-
-    created: datetime = models.DateTimeField(auto_now_add=True)
-
-    class Meta:
-        db_table = "recurring_reservation"
-        base_manager_name = "objects"
-
-    def __str__(self) -> str:
-        return f"{self.name}"
-
-    @property
-    def denied_reservations(self):
-        # Avoid a query to the database if we have fetched list already
-        if "reservations" in self._prefetched_objects_cache:
-            return [
-                reservation
-                for reservation in self.reservations.all()
-                if reservation.state == ReservationStateChoice.DENIED
-            ]
-
-        return self.reservations.filter(state=ReservationStateChoice.DENIED)
-
-    @property
-    def weekday_list(self):
-        if self.weekdays:
-            return [int(i) for i in self.weekdays.split(",")]
-        return []
+__all__ = [
+    "Reservation",
+]
 
 
 class Reservation(ExportModelOperationsMixin("reservation"), SerializableMixin, models.Model):
@@ -232,7 +101,7 @@ class Reservation(ExportModelOperationsMixin("reservation"), SerializableMixin, 
         default="",
     )
     home_city = models.ForeignKey(
-        City,
+        "applications.City",
         verbose_name=_("Home city"),
         related_name="home_city_reservation",
         on_delete=models.SET_NULL,
@@ -241,7 +110,7 @@ class Reservation(ExportModelOperationsMixin("reservation"), SerializableMixin, 
         help_text="Home city of the group or association",
     )
     age_group = models.ForeignKey(
-        AgeGroup,
+        "reservations.AgeGroup",
         verbose_name=_("Age group"),
         null=True,
         blank=True,
@@ -274,7 +143,7 @@ class Reservation(ExportModelOperationsMixin("reservation"), SerializableMixin, 
     priority = models.IntegerField(choices=PriorityChoice.choices, default=PriorityChoice.MEDIUM)
 
     user = models.ForeignKey(
-        User,
+        "users.User",
         null=True,
         verbose_name=_("User"),
         on_delete=models.SET_NULL,
@@ -286,10 +155,10 @@ class Reservation(ExportModelOperationsMixin("reservation"), SerializableMixin, 
     buffer_time_before = models.DurationField(verbose_name=_("Buffer time before"), blank=True, null=True)
     buffer_time_after = models.DurationField(verbose_name=_("Buffer time after"), blank=True, null=True)
 
-    reservation_unit = models.ManyToManyField(ReservationUnit, verbose_name=_("Reservation unit"))
+    reservation_unit = models.ManyToManyField("reservation_units.ReservationUnit", verbose_name=_("Reservation unit"))
 
     recurring_reservation = models.ForeignKey(
-        RecurringReservation,
+        "reservations.RecurringReservation",
         verbose_name=_("Recurring reservation"),
         related_name="reservations",
         on_delete=models.PROTECT,
@@ -300,7 +169,7 @@ class Reservation(ExportModelOperationsMixin("reservation"), SerializableMixin, 
     num_persons = models.fields.PositiveIntegerField(verbose_name=_("Number of persons"), null=True, blank=True)
 
     purpose = models.ForeignKey(
-        "ReservationPurpose",
+        "reservations.ReservationPurpose",
         verbose_name=_("Reservation purpose"),
         on_delete=models.SET_NULL,
         null=True,
@@ -308,7 +177,7 @@ class Reservation(ExportModelOperationsMixin("reservation"), SerializableMixin, 
     )
 
     cancel_reason = models.ForeignKey(
-        ReservationCancelReason,
+        "reservations.ReservationCancelReason",
         verbose_name=_("Reason for cancellation"),
         related_name="reservations",
         on_delete=models.PROTECT,
@@ -371,7 +240,7 @@ class Reservation(ExportModelOperationsMixin("reservation"), SerializableMixin, 
     )
 
     deny_reason = models.ForeignKey(
-        ReservationDenyReason,
+        "reservations.ReservationDenyReason",
         verbose_name=_("Reason for deny"),
         related_name="reservations",
         on_delete=models.PROTECT,
@@ -535,304 +404,6 @@ class Reservation(ExportModelOperationsMixin("reservation"), SerializableMixin, 
             f"{', '.join(unit_names)}\n"
             f"{self.reservation_unit.unit if hasattr(self.reservation_unit, 'unit') else ''}"
         )
-
-
-class ReservationPurpose(models.Model):
-    name = models.CharField(max_length=200)
-
-    class Meta:
-        db_table = "reservation_purpose"
-        base_manager_name = "objects"
-
-    def __str__(self) -> str:
-        return self.name
-
-
-class ReservationMetadataField(models.Model):
-    field_name = models.CharField(max_length=100, verbose_name=_("Field name"), unique=True)
-
-    class Meta:
-        db_table = "reservation_metadata_field"
-        base_manager_name = "objects"
-        verbose_name = _("Reservation metadata field")
-        verbose_name_plural = _("Reservation metadata fields")
-
-    def __str__(self) -> str:
-        return self.field_name
-
-
-class ReservationMetadataSet(models.Model):
-    name = models.CharField(max_length=100, verbose_name=_("Name"), unique=True)
-    supported_fields = models.ManyToManyField(
-        ReservationMetadataField,
-        verbose_name=_("Supported fields"),
-        related_name="metadata_sets_supported",
-    )
-    required_fields = models.ManyToManyField(
-        ReservationMetadataField,
-        verbose_name=_("Required fields"),
-        related_name="metadata_sets_required",
-        blank=True,
-    )
-
-    class Meta:
-        db_table = "reservation_metadata_set"
-        base_manager_name = "objects"
-        verbose_name = _("Reservation metadata set")
-        verbose_name_plural = _("Reservation metadata sets")
-
-    def __str__(self) -> str:
-        return self.name
-
-
-class ReservationStatistic(models.Model):
-    reservation = models.OneToOneField(Reservation, on_delete=models.SET_NULL, null=True)
-
-    reservation_created_at = models.DateTimeField(verbose_name=_("Created at"), null=True, default=timezone.now)
-
-    reservation_handled_at = models.DateTimeField(
-        verbose_name=_("Handled at"),
-        null=True,
-        blank=True,
-        help_text="When this reservation was handled.",
-    )
-
-    reservation_confirmed_at = models.DateTimeField(verbose_name=_("Confirmed at"), null=True)
-
-    buffer_time_before = models.DurationField(verbose_name=_("Buffer time before"), blank=True, null=True)
-    buffer_time_after = models.DurationField(verbose_name=_("Buffer time after"), blank=True, null=True)
-
-    updated_at = models.DateTimeField(verbose_name=_("Statistics updated at"), null=True, blank=True, auto_now=True)
-
-    reservee_type = models.CharField(
-        max_length=50,
-        choices=CustomerTypeChoice.choices,
-        null=True,
-        blank=True,
-        help_text="Type of reservee",
-    )
-
-    applying_for_free_of_charge = models.BooleanField(
-        verbose_name=_("Reservee is applying for a free-of-charge reservation"),
-        null=False,
-        default=False,
-        blank=True,
-    )
-
-    reservee_language = models.CharField(
-        verbose_name=_("Preferred language of reservee"),
-        max_length=255,
-        blank=True,
-        default="",
-    )
-
-    num_persons = models.fields.PositiveIntegerField(verbose_name=_("Number of persons"), null=True, blank=True)
-
-    priority = models.IntegerField(choices=PriorityChoice.choices, default=PriorityChoice.MEDIUM)
-
-    priority_name = models.CharField(max_length=255, null=False, default="", blank=True)
-
-    home_city = models.ForeignKey(
-        City,
-        verbose_name=_("Home city"),
-        related_name="reservation_statistics",
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-        help_text="Home city of the group or association",
-    )
-
-    home_city_name = models.CharField(
-        verbose_name=_("Home city name"),
-        max_length=100,
-        null=False,
-        default="",
-        blank=True,
-    )
-
-    home_city_municipality_code = models.CharField(
-        verbose_name=_("Home city municipality code"), default="", max_length=30
-    )
-
-    purpose = models.ForeignKey(
-        "ReservationPurpose",
-        verbose_name=_("Reservation purpose"),
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-    )
-
-    purpose_name = models.CharField(max_length=200, null=False, default="", blank=True)
-
-    age_group = models.ForeignKey(
-        AgeGroup,
-        verbose_name=_("Age group"),
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-    )
-
-    age_group_name = models.fields.CharField(max_length=255, null=False, default="", blank=True)
-
-    is_applied = models.BooleanField(
-        default=False,
-        blank=True,
-        verbose_name=_("Is the reservation done through application process."),
-    )
-
-    ability_group = models.ForeignKey(
-        AbilityGroup,
-        verbose_name=_("Ability group"),
-        null=True,
-        blank=True,
-        on_delete=models.SET_NULL,
-    )
-
-    ability_group_name = models.fields.TextField(
-        verbose_name=_("Name"),
-        null=False,
-        blank=False,
-    )
-
-    begin = models.DateTimeField(verbose_name=_("Begin time"))
-
-    end = models.DateTimeField(verbose_name=_("End time"))
-
-    duration_minutes = models.IntegerField(null=False, verbose_name=_("Reservation duration in minutes"))
-
-    reservation_type = models.CharField(
-        max_length=50,
-        null=True,
-        blank=False,
-        help_text="Type of reservation",
-    )
-
-    state = models.CharField(
-        max_length=32,
-        verbose_name=_("State"),
-    )
-
-    cancel_reason = models.ForeignKey(
-        ReservationCancelReason,
-        verbose_name=_("Reason for cancellation"),
-        related_name="reservation_statistics",
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-    )
-
-    cancel_reason_text = models.CharField(
-        max_length=255,
-        null=False,
-        blank=False,
-        verbose_name=_("The reason text of the cancel reason"),
-    )
-
-    deny_reason = models.ForeignKey(
-        ReservationDenyReason,
-        verbose_name=_("Reason for deny"),
-        related_name="reservation_statistics",
-        on_delete=models.SET_NULL,
-        null=True,
-        blank=True,
-    )
-
-    deny_reason_text = models.CharField(
-        max_length=255,
-        null=False,
-        blank=False,
-        verbose_name=_("The reason text of the deny reason"),
-    )
-
-    price = models.DecimalField(
-        verbose_name=_("Price"),
-        max_digits=10,
-        decimal_places=2,
-        default=0,
-        help_text="The price of this particular reservation",
-    )
-
-    price_net = models.DecimalField(
-        verbose_name=_("Price net"),
-        max_digits=20,
-        decimal_places=6,
-        default=0,
-        help_text="The price of this particular reservation excluding VAT",
-    )
-
-    non_subsidised_price = models.DecimalField(
-        verbose_name=_("Non subsidised price"),
-        max_digits=20,
-        decimal_places=2,
-        default=0,
-        help_text="The non subsidised price of the reservation excluding VAT",
-    )
-    non_subsidised_price_net = models.DecimalField(
-        verbose_name=_("Non subsidised net price"),
-        max_digits=20,
-        decimal_places=6,
-        default=0,
-        help_text="The non subsidised price of the reservation excluding VAT",
-    )
-    is_subsidised = models.BooleanField(help_text="Is the reservation price subsidised", default=False)
-    is_recurring = models.BooleanField(help_text="Is the reservation recurring", default=False)
-    recurrence_begin_date = models.DateField(verbose_name="Recurrence begin date", null=True)
-    recurrence_end_date = models.DateField(verbose_name="Recurrence end date", null=True)
-    recurrence_uuid = models.CharField(verbose_name="Recurrence UUID", max_length=255, default="", blank=True)
-    reservee_is_unregistered_association = models.BooleanField(
-        verbose_name=_("Reservee is an unregistered association"),
-        null=True,
-        default=False,
-        blank=True,
-    )
-    reservee_uuid = models.CharField(verbose_name="Reservee UUID", max_length=255, default="", blank=True)
-    tax_percentage_value = models.DecimalField(
-        verbose_name=_("Tax percentage value"),
-        max_digits=5,
-        decimal_places=2,
-        default=0,
-        help_text="The value of the tax percentage for this particular reservation",
-    )
-
-    primary_reservation_unit = models.ForeignKey(ReservationUnit, null=True, on_delete=models.SET_NULL)
-
-    primary_reservation_unit_name = models.CharField(verbose_name=_("Name"), max_length=255)
-    primary_unit_tprek_id = models.CharField(
-        verbose_name=_("TPREK id"),
-        max_length=255,
-        null=True,
-    )
-    primary_unit_name = models.CharField(verbose_name=_("Name"), max_length=255)
-
-    class Meta:
-        db_table = "reservation_statistics"
-        base_manager_name = "objects"
-
-    def __str__(self) -> str:
-        return f"{self.reservee_uuid} - {self.begin} - {self.end}"
-
-
-class ReservationStatisticsReservationUnit(models.Model):
-    reservation_statistics = models.ForeignKey(
-        ReservationStatistic,
-        on_delete=models.CASCADE,
-        related_name="reservation_stats_reservation_units",
-    )
-    reservation_unit = models.ForeignKey(ReservationUnit, null=True, on_delete=models.SET_NULL)
-    unit_tprek_id = models.CharField(
-        verbose_name=_("TPREK id"),
-        max_length=255,
-        null=True,
-    )
-    name = models.CharField(verbose_name=_("Name"), max_length=255)
-    unit_name = models.CharField(verbose_name=_("Name"), max_length=255)
-
-    class Meta:
-        db_table = "reservation_statistics_reservation_unit"
-        base_manager_name = "objects"
-
-    def __str__(self) -> str:
-        return f"{self.reservation_statistics} - {self.reservation_unit}"
 
 
 AuditLogger.register(Reservation)
